+++ src/sdk/Utils.ts
@@ -32,14 +32,14 @@ export function buildEvmDataCell(transactionLinker: TransactionLinker, evmProxyM
         : null;
 
     const json = JSON.stringify({
-        evm_call: {
+        evmCall: {
             target: evmProxyMsg.evmTargetAddress,
-            method_name: formatSolidityMethodName(evmProxyMsg.methodName),
+            methodName: formatSolidityMethodName(evmProxyMsg.methodName),
             arguments: evmArguments,
-            gas_limit: Number(evmProxyMsg.gasLimit)
+            gasLimit: Number(evmProxyMsg.gasLimit),
         },
-        shards_key: transactionLinker.shardsKey,
-        shard_count: transactionLinker.shardCount,
+        shardsKey: transactionLinker.shardsKey,
+        shardCount: transactionLinker.shardCount,
     });
 
     return beginCell().storeStringTail(json).endCell();